+++ src/Components/MyInfo/src/stores/index.ts
@@ -5,7 +5,7 @@ configure({
   enforceActions: 'observed',
 })
 
-export interface IPingItem {
+export interface PingItemProps {
   time: number
 }
 
@@ -14,7 +14,7 @@ class MyInfoStore {
   public readonly conf = conf[this.ID]
 
   @observable public isPing: boolean = false
-  @observable public pingItems: IPingItem[] = []
+  @observable public pingItems: PingItemProps[] = []
 
   @observable public refs = {}
 
@@ -29,7 +29,7 @@ class MyInfoStore {
   }
 
   @action
-  public setPingItems = (pingItems: IPingItem[]) => {
+  public setPingItems = (pingItems: PingItemProps[]) => {
     this.pingItems = pingItems
   }
 
@@ -39,7 +39,7 @@ class MyInfoStore {
   }
 
   @action
-  public appendPingItem = (item: IPingItem) => {
+  public appendPingItem = (item: PingItemProps) => {
     this.pingItems.push(item)
   }
 }